requirements.txt
streamlit
pandas
pyyaml
altair
matplotlib
numpy